% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{check.dag}
\alias{check.dag}
\title{Internal check of whether the formulas passed to the function dpa follow a directed acyclic graph (DAG)}
\usage{
check.dag(meta, data)
}
\arguments{
\item{meta}{meta data obtained internally inside the dpasurv::dpa function}

\item{data}{input data to the dpasurv::dpa function}
}
\value{
this function doesn't return anything
}
\description{
Do not call this function on its own
}
\examples{
library(dpasurv)

data(simdata)

meta <- get.meta(Surv(start, stop, event) ~ x + M, M ~ x, simdata)
check.dag(meta, simdata)
}
\keyword{internal}
