% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpa.R
\name{dpa}
\alias{dpa}
\title{Dynamic Path Analysis}
\usage{
dpa(
  out.formula,
  mediator.formulas,
  id,
  data,
  boot.n = 100,
  progress_bar = FALSE,
  ...
)
}
\arguments{
\item{out.formula}{Survival formula for Aalen's additive hazards model.}

\item{mediator.formulas}{Mediator regression formula (in case of a single mediator), or a list of regression formulas (in case of multiple mediators).
The formulas must be ordered according to Directed Acyclic Graph Structure (see Details).}

\item{id}{character string indicating which column of 'data' corresponds to the subject ID.}

\item{data}{Data set in counting process format. In particular the data should contain a "start", "stop" and "event" column along with
any mediators and baseline covariates.}

\item{boot.n}{Number of bootstrap samples.}

\item{progress_bar}{Boolean. If TRUE, show progress bar. Defaults to FALSE.}

\item{...}{other parameters passed to Aalen's additive hazards regression function "timereg::aalen()"}
}
\value{
Object of class `dpa` with following fields:
\describe{
  \item{coefs}{list of estimated coefficients from each of the regressions listed in out.formula and mediator.formulas.}
  \item{boot.coefs}{list of bootstrap estimates corresponding to coefs. This stores all the bootstrap estimates to facilitate
calculation of direct, indirect and total effects along with bootstrap confidence intervals.}
  \item{meta}{a list keeping track of responses and covariates of each of the out.formula and mediator.formulas. Also keeps
track of all variable types and level names in case of factors.}
}
}
\description{
Dynamic Path Analysis
}
\details{
\code{dpa} performs Dynamic Path Analysis of a Directed Acyclic Graph. The out.formula
can have as covariates all mediators listed in mediator.formulas. The mediator.formulas must obey the
following DAG structure rule: The response of the k-th formula cannot appear as covariate in any of the formulas
k+1, ..., length(mediator.formulas).
}
\examples{
library(dpasurv)

data(simdata)

s <- dpa(survival::Surv(start,stop,event)~M+x, list(M~x), id="subject", data=simdata, boot.n=100)

direct <- effect(x ~ outcome, s)
indirect <- effect(x ~ M ~ outcome, s)
total <- sum(direct, indirect)

par(mfrow=c(1,3))
plot(direct); abline(h=0, lty=2, col=2)
plot(indirect); abline(h=0, lty=2, col=2)
plot(total); abline(h=0, lty=2, col=2)

# Multiple treatment arms:
s2<-dpa(survival::Surv(start,stop,event)~M+dose,list(M~dose),id="subject",data=simdata,boot.n=100)

direct2 <- effect(dose ~ outcome, s2)
indirect2 <- effect(dose ~ M ~ outcome, s2)
total2 <- sum(direct2, indirect2)

par(mfrow=c(2,3))
plot(direct2)
plot(indirect2)
plot(total2)

}
