% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{find.variables}
\alias{find.variables}
\title{Find variables of a given formula}
\usage{
find.variables(formula)
}
\arguments{
\item{formula}{}
}
\value{
character vector of the formula's variable names
}
\description{
Do not call this function on its own
}
\keyword{internal}
