% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{resolve.ties}
\alias{resolve.ties}
\title{Resolves ties in event times by adding a small random noise to the duplicates}
\usage{
resolve.ties(stopt, event, obstimes)
}
\arguments{
\item{stopt}{stop times of the (start,stop] intervals across all subjects and time points}

\item{event}{event status corresponding to each stopt}

\item{obstimes}{sorted observed event times across all subjects}
}
\value{
stopt with duplicates adjusted by the small random noise. If no ties, then the function just returns stopt.
}
\description{
This function mimics how ties are handled in timereg::read.surv()
}
