% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{Surv}
\alias{Surv}
\title{A wrapper function for survival::Surv}
\usage{
Surv(...)
}
\arguments{
\item{...}{parameters passed to survival::Surv}
}
\value{
object of class survival::Surv
}
\description{
A wrapper function for survival::Surv
}
\examples{
library(dpasurv)

data(simdata)

survival.obj <- Surv(simdata$start, simdata$stop, simdata$event)
}
\keyword{internal}
