% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{add.ci}
\alias{add.ci}
\title{Calculate bootstrap confidence bands for the direct or indirect effect of interest}
\usage{
add.ci(object, alpha)
}
\arguments{
\item{object}{direct or indirect effect of interest (as obtained within a call to the function \code{effect}).}

\item{alpha}{the confidence level}
}
\value{
data.frame containing the unique event times, estimated effect, and lower and upper confidence bands
}
\description{
Calculate bootstrap confidence bands for the direct or indirect effect of interest
}
\keyword{internal}
