% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get.meta}
\alias{get.meta}
\title{Internal function that gathers meta data on the input to function dpa}
\usage{
get.meta(out.formula, mediator.formulas, data)
}
\arguments{
\item{out.formula}{obtained directly from corresponding input to dpasurv::dpa}

\item{mediator.formulas}{obtained directly from corresponding input to dpasurv::dpa}

\item{data}{obtained directly from corresponding input to dpasurv::dpa}
}
\value{
this function returns meta data associated with the call to dpasurv::dpa
}
\description{
Do not call this function on its own
}
\examples{
library(dpasurv)

data(simdata)

meta <- get.meta(Surv(start, stop, event) ~ x + M, M ~ x, simdata)
}
\keyword{internal}
