% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.effect}
\alias{plot.effect}
\title{Plot effects from dynamic path analysis along with bootstrap confidence bands}
\usage{
\method{plot}{effect}(x, relative = FALSE, ...)
}
\arguments{
\item{x}{object of class "effect"}

\item{relative}{should the effect be plotted on a relative survival scale (i.e. `y=exp(-effect)`)?. Defaults to FALSE.}

\item{...}{other graphical parameters passed to the graphics::plot function.}
}
\description{
plotting method for class "effect"
}
\examples{
library(dpasurv)

data(simdata)

set.seed(1)

s <- dpa(survival::Surv(start,stop,event)~M+x, list(M~x), id="subject", data=simdata, boot.n=50)

direct <- effect(x ~ outcome, s)
indirect <- effect(x ~ M ~ outcome, s)
total <- sum(direct, indirect)

par(mfrow=c(1,3))
plot(direct); abline(h=0, lty=2, col=2)
plot(indirect); abline(h=0, lty=2, col=2)
plot(total); abline(h=0, lty=2, col=2)

}
