% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{Surv}
\alias{Surv}
\title{Wrapper function for survival::Surv}
\usage{
Surv(...)
}
\arguments{
\item{...}{parameters passed to survival::Surv}
}
\value{
object of class survival::Surv
}
\description{
Wrapper function for survival::Surv
}
\keyword{internal}
